Update Docker version and cron schedule
The Docker version is updated to 7.1.0 in the requirements.txt file. Additionally, the scheduling cron job in the Docker workflow (docker.yml) has been modified to run at the beginning of every month. The changes ensure that the system uses an updated version and the docker workflow follows a monthly schedule.